Refactor dumper
